Move viewmodels to the separate namespaces
