+++ GieldaL2.API/Controllers/OffersController.cs
@@ -1,5 +1,7 @@
 ﻿using System.Collections.Generic;
 using GieldaL2.API.ViewModels;
+using GieldaL2.API.ViewModels.Edit;
+using GieldaL2.API.ViewModels.View;
 using Microsoft.AspNetCore.Mvc;
 
 namespace GieldaL2.API.Controllers
