+++ GieldaL2.API/Controllers/TransactionsController.cs
@@ -1,5 +1,6 @@
 ﻿using System.Collections.Generic;
 using GieldaL2.API.ViewModels;
+using GieldaL2.API.ViewModels.View;
 using Microsoft.AspNetCore.Mvc;
 
 namespace GieldaL2.API.Controllers
