+++ GieldaL2.API/ViewModels/Edit/EditBuyOfferViewModel.cs
@@ -1,14 +1,14 @@
 ﻿using System.ComponentModel.DataAnnotations;
 
-namespace GieldaL2.API.ViewModels
+namespace GieldaL2.API.ViewModels.Edit
 {
-    public class EditSellOfferViewModel
+    public class EditBuyOfferViewModel
     {
         [Required]
         public int UserId { get; set; }
 
         [Required]
-        public int ShareId { get; set; }
+        public int StockId { get; set; }
 
         [Required]
         public int Amount { get; set; }
