+++ GieldaL2.API/ViewModels/Edit/EditShareViewModel.cs
@@ -1,7 +1,7 @@
 ﻿using System.ComponentModel;
 using System.ComponentModel.DataAnnotations;
 
-namespace GieldaL2.API.ViewModels
+namespace GieldaL2.API.ViewModels.Edit
 {
     public class EditShareViewModel
     {
