+++ GieldaL2.API/ViewModels/Edit/EditStockViewModel.cs
@@ -1,6 +1,6 @@
 ﻿using System.ComponentModel.DataAnnotations;
 
-namespace GieldaL2.API.ViewModels
+namespace GieldaL2.API.ViewModels.Edit
 {
     public class EditStockViewModel
     {
