+++ GieldaL2.API/ViewModels/View/StockViewModel.cs
@@ -1,6 +1,4 @@
-﻿using System.Collections.Generic;
-
-namespace GieldaL2.API.ViewModels
+﻿namespace GieldaL2.API.ViewModels.View
 {
     public class StockViewModel
     {
