+++ GieldaL2.API/ViewModels/View/TransactionViewModel.cs
@@ -1,9 +1,10 @@
-﻿namespace GieldaL2.API.ViewModels
+﻿namespace GieldaL2.API.ViewModels.View
 {
-    public class BuyOfferViewModel
+    public class TransactionViewModel
     {
         public int Id { get; set; }
-        public int UserId { get; set; }
+        public int BuyerId { get; set; }
+        public int SellerId { get; set; }
         public int StockId { get; set; }
         public int Amount { get; set; }
         public decimal Price { get; set; }
