+++ GieldaL2.API/ViewModels/View/UserViewModel.cs
@@ -1,6 +1,4 @@
-﻿using System.Collections.Generic;
-
-namespace GieldaL2.API.ViewModels
+﻿namespace GieldaL2.API.ViewModels.View
 {
     public class UserViewModel
     {
